pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dbdie_ml"
authors = [{ name = "trOOnies", email = "example@github.com" }]
description = "DBD Info Extraction ML Package"
readme = "README.md"
requires-python = ">=3.7.16"
keywords = ["dbd", "ml", "pytorch"]
license = { text = "GNU-3" }
classifiers = ["Framework :: Pytorch", "Programming Language :: Python :: 3"]
dependencies = [
    "numpy==1.21.5",
    "Pillow==9.4.0",
    "pydantic==2.5.3",
    "pydantic_core==2.14.6",
    "python-dotenv==0.21.1",
    "PyYAML==6.0.1",
    "torch==2.2.0+cu121",
    "torchsummary==1.5.1",
    "torchvision==0.17.0+cu121",
    'importlib-metadata; python_version<"3.10"',
]
dynamic = ["version"]

# [project.optional-dependencies]
# pdf = ["ReportLab>=1.2", "RXP"]
# rest = ["docutils>=0.3", "pack ==1.1, ==1.3"]

[project.scripts]
my-script = "dbdie_ml.module:function"
